fix aes code to be compatible with rust 1.82+
In Rust 1.82 they now use more of the Zkn extensions, most
of which are not implemented in this CPU. In particular, when
doing a copy, it uses the "PACK" instruction to accelerate
the operation, which we don't have.

This refactor eliminates any code except for the in-line
assembly in the regions that have Zkn enabled, which prevents
the generation of instructions that we can't handle.