add rram disturb test
this test will check if the RRAM written values were "disturbed".
It's meant to be run at the end of all the other tests, and helps
to catch if e.g. the write had some failure to make the data
persistent. In particular, we want to rule out that the verification
that happened immediately after the write was just reading back
from the data cache. By running this test at the end of all the
other tests in the pipeline, we ensure that we are probabilistically
running out of a "cold cache" with respect to the RRAM values.