re-order BIO-tests to put DMA first
Allows us to check DMA function first, which is the hardest
corner case to pass.

This also includes the work-around to initialize all the cores
using 'multiverse' to prevent random data in the program memory
from jamming the DMA bus.

More effort is needed to figure out the exact set of restrictions
to get DMA reliable with the chip's timing as taped-out.